check if requested file exists before trying to copy it

+++ modules/repos.py
@@ -54,7 +54,8 @@ class YUMRepos:
 
             if not self.override:
                 os.makedirs('/tmp' + self.path, exist_ok=True)
-                shutil.copyfile(filename, '/tmp' + filename)
+                if os.path.isfile(filename):
+                    shutil.copyfile(filename, '/tmp' + filename)
 
             with open(filename, 'w') as f:
                 f.write(content)
